+++ packages/theme-sdk/src/index.ts
@@ -6,4 +6,5 @@ export * from "./pages";
 export * from "./routing";
 export * from "./sidebar";
 export * from "./storage";
+export * from "./system";
 export * from "./usage";